Move guild scanning queue logic to GuildScanner
